refactor: controller interface 추가
