+++ app/src/main/java/nextstep/jwp/http/RequestHandler.java
@@ -1,6 +1,9 @@
 package nextstep.jwp.http;
 
-import nextstep.jwp.controller.*;
+import nextstep.jwp.controller.Controller;
+import nextstep.jwp.controller.LoginController;
+import nextstep.jwp.controller.RegisterController;
+import nextstep.jwp.controller.StaticResourceController;
 import nextstep.jwp.exception.NoMatchingControllerException;
 import nextstep.jwp.exception.ResourceNotFoundException;
 import nextstep.jwp.http.request.HttpRequest;
